+++ client/package.json
@@ -15,6 +15,7 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router-dom": "^6.3.0",
+    "react-tinder-card": "^1.4.5",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
