+++ client/src/components/Header/Header.css
@@ -2,7 +2,7 @@
     display: flex;
     justify-content:space-around;
     align-items: center;
-   padding: 15px;
+   padding: 20px;
 }
 .logo{
     height: 60px;
